requirements.txt
requests
nltk
pandas
python-dotenv
shiny